+++ examples/Realtime/realtimeClient.php
@@ -55,6 +55,12 @@ $rtc->on('live-stopped', function (\InstagramAPI\Realtime\Payload\LiveBroadcast 
 $rtc->on('direct-story-created', function (\InstagramAPI\Response\Model\DirectThread $thread) {
     printf('[RTC] Story %s has been created%s', $thread->getThreadId(), PHP_EOL);
 });
+$rtc->on('thread-item-liked', function ($threadId, $threadItemId, \InstagramAPI\Response\Model\DirectThreadItem $threadItem) {
+    printf('[RTC] Item %s has been liked in %s%s', $threadItemId, $threadId, PHP_EOL);
+});
+$rtc->on('thread-item-unliked', function ($threadId, $threadItemId, \InstagramAPI\Response\Model\DirectThreadItem $threadItem) {
+    printf('[RTC] Item %s has been unliked in %s%s', $threadItemId, $threadId, PHP_EOL);
+});
 $rtc->on('direct-story-updated', function ($threadId, $threadItemId, \InstagramAPI\Response\Model\DirectThreadItem $threadItem) {
     printf('[RTC] Item %s has been created in story %s%s', $threadItemId, $threadId, PHP_EOL);
 });